building with gcc now results in a working library
when parsing the json file, after completely populating the screen
structure with all of the child structures specified in the json
the code would attempt to populate the screens parent with the
screeb structure.  This is kind of silly because the screen is the
root of the user interface and has no parent.

the real question here is how in the name of keyush the stunt dog
did this ever work when compiled with clang? (p.s. it worked fine!)

+++ src/color.c
@@ -112,9 +112,9 @@ static INLINE void do_set_bg(void)
     // are we setting a rgb background?
     if(attrs.bytes[ATTR] & BG_RGB)
     {
-        params[0] = (uint8_t)attrs.bytes[BG_R];
-        params[1] = (uint8_t)attrs.bytes[BG_G];
-        params[2] = (uint8_t)attrs.bytes[BG_B];
+        params[0] = attrs.bytes[BG_R];
+        params[1] = attrs.bytes[BG_G];
+        params[2] = attrs.bytes[BG_B];
 
         f_str = &rgb_seq[0];
         parse_format();
